Implement Flight#departure_airport_code method
This protects us from breaking the law of demeter, meaning that we're
not coupled to a specific implementation. By calling
flight.departure_airport.code in our views, we are coupled to that
specific implementation. What if we want to change what that association
is called, or maybe it is used in another class? When we break the law
of demeter, we're unable to refactor the implementation w/o breaking
lots of stuff in our views.

The Flight#departure_airport_code method serves as an *interface* in
this particular case. Anywhere we need the airport code of the departure
airport, we use this method. If we need to change how this is
imlemented, now we only need to do it in one place.

Is it more code? Yes. But I think the other wins far outweigh three
extra lines of code (plus tests).